+++ backend/pokerogue/src/main/java/com/pokerogue/helper/ability/dto/AbilityPokemonResponse.java
@@ -1,0 +1,12 @@
+package com.pokerogue.helper.ability.dto;
+
+import java.util.List;
+
+public record AbilityPokemonResponse(
+        String id,
+        Long pokedexNumber,
+        String koName,
+        String image,
+        List<AbilityTypeResponse> pokemonTypeResponses
+) {
+}
