+++ backend/pokerogue/src/main/java/com/pokerogue/helper/ability/dto/AbilityTypeResponse.java
@@ -1,0 +1,4 @@
+package com.pokerogue.helper.ability.dto;
+
+public record AbilityTypeResponse(String typeLogo, String typeName) {
+}
