--- backend/pokerogue/src/main/java/com/pokerogue/helper/ability2/dto/AbilityTypeResponse.java
@@ -1,4 +1,0 @@
-package com.pokerogue.helper.ability2.dto;
-
-public record AbilityTypeResponse(String typeLogo, String typeName) {
-}
