+++ backend/pokerogue/src/main/java/com/pokerogue/helper/pokemon/data/Evolution.java
@@ -1,0 +1,5 @@
+package com.pokerogue.helper.pokemon.data;
+
+public record Evolution(String from, String level, String to, String item, String condition) {
+
+}
