+++ backend/pokerogue/src/main/java/com/pokerogue/helper/pokemon/dto/EvolutionResponse.java
@@ -1,0 +1,14 @@
+package com.pokerogue.helper.pokemon.dto;
+
+
+public record EvolutionResponse(
+        String id,
+        String name,
+        Integer level,
+        Integer depth,
+        String item,
+        String condition,
+        String image
+) {
+
+}
