+++ backend/pokerogue/src/main/java/com/pokerogue/helper/pokemon/dto/EvolutionResponses.java
@@ -1,0 +1,7 @@
+package com.pokerogue.helper.pokemon.dto;
+
+import java.util.List;
+
+public record EvolutionResponses(int currentDepth, List<EvolutionResponse> stages) {
+
+}
