--- backend/pokerogue/src/main/java/com/pokerogue/helper/pokemon2/data/Evolution.java
@@ -1,5 +1,0 @@
-package com.pokerogue.helper.pokemon2.data;
-
-public record Evolution(String from, String level, String to, String item, String condition) {
-
-}
