--- backend/pokerogue/src/main/java/com/pokerogue/helper/pokemon2/dto/EvolutionResponses.java
@@ -1,7 +1,0 @@
-package com.pokerogue.helper.pokemon2.dto;
-
-import java.util.List;
-
-public record EvolutionResponses(int currentDepth, List<EvolutionResponse> stages) {
-
-}
