--- backend/pokerogue/src/main/java/com/pokerogue/helper/pokemon2/dto/PokemonAbilityResponse.java
@@ -1,8 +1,0 @@
-package com.pokerogue.helper.pokemon2.dto;
-
-public record PokemonAbilityResponse(String id, String name, String description, Boolean passive, Boolean hidden) {
-
-    public static PokemonAbilityResponse from(String name, String description, boolean passive, boolean hidden) {
-        return new PokemonAbilityResponse(name, name, description, passive, hidden);
-    }
-}
