+++ backend/pokerogue/src/main/java/com/pokerogue/helper/type/dto/PokemonTypeResponse.java
@@ -1,6 +1,5 @@
 package com.pokerogue.helper.type.dto;
 
-import com.pokerogue.helper.pokemon2.data.Pokemon;
 import com.pokerogue.helper.type.domain.PokemonType;
 
 public record PokemonTypeResponse(String typeName, String typeLogo) {
